LayoutTests/fast/canvas/webgl/webgl2-buffers-expected.txt

CONSOLE MESSAGE: line 30: TypeError: null is not an object (evaluating 'context.ARRAY_BUFFER')
Make sure that reading and writing WebGL2 buffers works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
